update MessageBox Window

+++ ChatGPT/MainWindow.xaml.cs
@@ -139,7 +139,12 @@ namespace ChatAI
             // Access the selected chat
             var selectedChat = ViewModel.SelectedChat;
             // Delete the selected chat
-            ViewModel.DeleteChat(selectedChat);
+            MessageBoxCustom result = new MessageBoxCustom("Notification", "Are you sure, You want to delete this chat?", MessageType.Info, MessageButtons.YesNo);
+            result.ShowDialog();
+            if (result.DialogResult == true)
+            {
+                ViewModel.DeleteChat(selectedChat);
+            }
             //ViewModel.NewChat();
         }
 
@@ -153,7 +158,8 @@ namespace ChatAI
             String getTopic = AddTopic.Text;
             if (String.IsNullOrWhiteSpace(getTopic))
             {
-                System.Windows.MessageBox.Show("Please fill topic box!", "Notification", MessageBoxButton.OK, MessageBoxImage.Warning);
+                MessageBoxCustom mb = new MessageBoxCustom("Notification", "Please fill topic box!", MessageType.Warning, MessageButtons.OK);
+                mb.ShowDialog();
             }
             else
             {
@@ -170,7 +176,8 @@ namespace ChatAI
                     };
                     comboBoxItems.Insert(newItemModel);
                 }
-                System.Windows.MessageBox.Show("Add your topic successful", "Notification", MessageBoxButton.OK, MessageBoxImage.Information);
+                MessageBoxCustom mb = new MessageBoxCustom("Notification", "Add your topic successful!", MessageType.Success, MessageButtons.OK);
+                mb.ShowDialog();
             }
         }
